Refactor Makefiles
Do not create output directory automatically as this appears to
confuse Make and results in it trying to run recipes when there
have been no changes. Instead switch to having an 'empty' output
directory (with a .gitkeep as Git does not allow truly empty
directories to be added) which means output will exist as part
of a clone.